resize/requirements.txt
Pillow>=10.0.0
pillow-avif-plugin>=1.3.0 